feat : rebuild component
